refactor: move code to header.tsx
